+++ bootstrap.js
@@ -1,3 +1,7 @@
+if (alchemy.plugins.form) {
+	throw new Error('The alchemy-form plugin has to be loaded AFTER alchemy-widget');
+}
+
 Router.add({
 	name       : 'AlchemyWidgets#save',
 	methods    : 'post',