+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Allow hiding widgets from the add-menu
 * Let actions return their button contents as elements instead of only html
 * Cancel clicks on widgets when editing them
+* Fix getting the `hawkejs_renderer` instance in a widget
 
 ## 0.1.5 (2022-07-14)