+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Use `alchemy-chimera` style for the widget configuration dialog
 * Make renders wait for widgets that have to render their content asynchronously
 * Allow setting the element to use in a Text widget
+* Fix Header-widget level actions
 
 ## 0.1.5 (2022-07-14)