+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Wait for widgets to render their contents before starting editor
 * Add `can_be_removed` property to widget elements
 * Add `can_be_moved` property to widget elements
+* Throw an error if `alchemy-form` is loaded before this plugin
 
 ## 0.1.5 (2022-07-14)